direct-access: Fix write handler when there is little padding after the sector.
Firstly, round up the producr index at the end of the write. This way we will
process every bit sent to us while the write-gate is asserted.

Secondly, we only expect to see 2*4489 because of the way to we sync to that
pattern in IRQ_wdata_dma(). Hence the header is one word shorter than we might
naively expect, and we must adjust the arithmetic to account for the shortfall.

This fixes the "Write Issue?" failure seen in the file selector on Atari ST.