floppy: Ensure that RDATA pin is deasserted when drive is not selected.
Specifically, the gpio_configure_pin() in rdata_start() sets the pin
default state to O_TRUE (0) if the drive is selected when the read
stream starts. This causes the pin to be erroneously asserted in
_IRQ_SELA_changed when the drive is deselected.

One effect of this happened to be that drive A is drowned out on a
CPC 6128 if F-F was connected as drive B. The CPC 6128 selects drive B
most of the time (since SELB is generated simply as !SELA) and hence
made this bug extremely likely to trigger.

Thanks to ikonsgr on EAB (and Ebay!) for the bug report. His
Spectrum/Amstrad drive cables will now work properly with FlashFloppy!