Add support for PC98 host, with raw and FDI formats.
For the img format, host must be set to pc98 in order to get the
most common 360 rpm formats.

For the FDI format (from the Anex86 emulator, not to be confused
with the Amiga FDI format), parameters are loaded from the image
header.

I have not tested anything other than 1.2MB as I don't have any
disks or images of other types.