i2c: Revert 07953affa5e1f and implement correct fix.
The last byte of a multi-DMA i2c transfer could be lost, not because
of an error in the i2c display controller, but because of a bug in
FlashFloppy's I2C DMA handling.

When the last byte of a transmit is acknowledged by the receiver,
I2C sets the BTF flag in SR1. This is not cleared until SR1 is read
and a further byte is transmitted. Unfortunately this means that if
we are slow to start a new DMA after the previous completes, BTF
gets set. And it won't clear because we don't read SR1. Then at the
end of the final DMA, we wait for BTF and then issue STOP. But
because BTF is already set, we issue STOP too early and the final
byte is not transmitted.

The fix is to always wait for BTF at the end of a DMA transfer.
Because this requires us to wait for the final two bytes (one in the
data register, and one in the shift register) the delay can be 50us
(Fast I2C) up to 200us (Standard I2C). We avoid spin-waiting by
enabling the I2C Event IRQ and calling back into the DMA pipeline
when we are interrupted on BTF.