Fix quickdisk now that image_setup_track() is void
In 1937c17 responsibility for entering and exiting D-A was removed from
image.c and moved into the floppy handling. With that change
image_setup_track no longer need a return value. Since quickdisk doesn't
support D-A, the change is as simple as removing the condition after
running image_setup_track.