ringio: Fix writing faults caused by seeking backward
There may be pending writes past wd_prod because seeking can rewind
cursors into the recent past. Assertions were only triggering if the
track was fully buffered, but I believe that was because the writes were
just completing faster in that case and so were more likely to 'catch
up' to the rewound producer.