img: Properly seek on track setup
The old code was horribly broken, but the bug fixed in a505765 make it
harder to notice as it would write out garbage in a correct-looking way.
Only the first sector or so would be messed up and the reader generally
isn't looking at the sector so soon after the seek anyway. With the fix
in a505765 the garbage no longer looked correct so made this bug more
apparent.

The previous code is broken in many ways, like not handling the bc
cursors properly on roll-over and having the consumer ahead of the
producer even though bc_rdata_flux doesn't support that. All of that was
supposed to be temporary to get things into a half-way working shape and
should not have been included in d8b0b2a, but was forgotten about after
resolving other issues and it not causing an obvious problem.