img: Fix reads for sectors greater than 512 bytes
This reduces the batch size from 1024 to 256. There's no real harm to a
batch size smaller than 1024 because the batch size is no longer tied to
I/O size. 512 bytes would work, but requires half of the 2 KB read_bc to
be empty. 256 bytes ekes out 50% more of read_bc for buffering while
still being large enough for efficient buffer filling.

The problem with 1024 bytes is that bc_rdata_flux may not drain all of
read_bc. bc_rdata_flux only reads from read_bc when there are at least 4
bytes. Since odd numbers of data bytes are used for many structures, it
is trivial for 2 bc bytes to get "stuck" in read_bc and cannot drain
until further bitcells follow it. While that could be fixed, it is
better to not require the entire read_bc being empty, as that leaves a
strong risk of rdata underrun.