Drain ring_io I/O when changing tracks and image sync
Rapid seeking can cause a backlog of reads to be queued. This doesn't
cause too much trouble because fs_async swaps to synchronous processing
when the queue is full. However, the queued reads do have an impact on
latency once the final track is reached. This latency is noticeable, but
could be quite dramatic if the fs_async queue size were increased.

When fs_async has to block it logs, so avoiding the blocking is also
helpful to avoid polluting logs.

d2c1990 increased the fs_async queue size to 8 because a single seek
could have exceeded the 4-entry queue (4 for old track, 4 more for new
track). With the draining in place the number of I/O ops is strictly
limited to 4, so the queue size can go back to its earlier value of 4.

I/O processing after the image handler is done would have continued
writing to read_data. This was not a problem for dummy_handler and when
floppy_cancel() was next called, which were the only cases present.
Draining the I/O prevents this gotcha from biting anyone as the code
changes.