ringio: Allow cached reads to be non-contiguous
This is very necessary to support interleaved blocks, as we simply can't
wait for the read producer to read up to half the track just to reach
the second physical sector.

It is also helpful for skewed sides, when the head is changed
immediately after a track step. In that case the "wrong" side is chosen
first and a skew may cause the read to go "backward".