img: Decrease chance of RDATA underrun
It now seeks to next sector sooner in the ring_io to give more time to
read the soon-to-be-needed blocks if sectors are interleaved.

For ED rates, disable the shadow ring. In addition use the same batch
size as used before ring_io (primarily to benefit ED, but applied for
all rates). It seems strange that a batch size of 1 was necessary for ED
which we'd think would benefit from larger batch sizes, but both changes
were necessary to avoid buffer underruns for the 2m88 FF_TestBed test
image.