ringio: Advance rd->prod immediately on seek
The previous intention was that enqueue_io would advance the rd->prod
cursor. However, that didn't actually happen because enqueue_io
short-circuits if fop_cb != NULL. We could move the short-circuit after
the cursor updates, but that would cause the recalculations every call
to ring_io_progress. So instead, copy the rd->prod advancement code to
ring_io_seek.

This explains why we needed such a small batch size for 2m88 in 282c4aa,
as an I/O op actually had to complete to make any reads available. With
this bug fix, batch sizes of 2 and 4 now work again for IMG. I used 2 to
be more interleaving-friendly.