adf: Use custom write buffer for "open loop" writes
This reverts part of the write code from 07d5b9e and replaces it with a
custom write buffer. Reads do not _block_ on the write buffer but are
delayed until it is drained. This lets us use all async code paths while
also supporting lots of writes without the controller doing any reads or
index synchronization.