adf: Fix bug that re-wrote blocks
Already-written blocks were re-written because the local 'c' variable
that mirrors cons wasn't updated. This wouldn't _create_ any corruptions
as either the data was just old, so re-write just-written blocks, or the
producer is close to wrapping-around the ring, so early-write new data.
However, if there was a CRC failure it could cause that block to be
written to disk when ordinarily it would not be written.