adf: Remove mis-placed ring_io_seek
In a8ae61f writing no longer uses ring_io, however a ring_io_seek was
left present in the write path. It seems like that should cause havoc,
but I suspect in the basic testing either the ring was already
fully-buffered (such that there were no further reads) or that the
reader was able to keep ahead of the writer until ring io was
fully-buffered.