da: Convert to async I/O
This uses a custom write buffer similar to ADF. Unlike ADF though, the
read caching is preserved when writing. This is partly because the
default track size is small enough (4KB) that we can just pay the cost
all the time. But it also avoids the need to wait for a pending read to
complete to make memory available for processing written commands.
However, if the D-A controller extends the track size past 8 sectors
with CMD_SET_LBA, then the read cache becomes useless.