Hard-code async write_bc.len
Previously write_bc's size was calculated to make sure writes for an HD
track in HFE could fit in memory. But that calculation is only necessary
when 1) there is enough memory to make such a thing possible (>30kB,
with min 10kB read_data) but 2) there's not enough memory to make it
"just work" (<53kB). Since 10+kB of memory is dedicated to other tasks,
this really only leaves a SoC RAM size of 48kB, yet we don't target
48kB.

Swapping to hard-code+ASSERT is simpler and avoids changing behavior
between 32 and 64 kB RAM sizes. It also avoids HFE-oriented logic in
floppy_generic.c.