dsk,img: Avoid completing writes when they overrun reads
If reading from flash is delayed, WDATA may overtake it and require
pausing WDATA processing. Previously if this happened at the end of the
write, the tail of the write would be lost. This seems unlikely, except
maybe short writes very soon after changing tracks. I was not able to
trigger this case.

ADF was already mostly correct, excepting the sec_idx!=nr_secs check.
But clearing the flag is clearer and mirrors other image formats.