hfe: Handle wraparound during writing
Seems this was accidentally lost with the conversion to async. This
probably doesn't matter much, but for v3 the index pulse could be wildly
misaligned and then this could become noticable.

The lack of wraparound was observable when using a logic analyzer doing
some North Star testing, but it didn't cause issues because it just lost
the tail end of the hard sector which contains no data. I only just
realized that the wraparound logic was previously present.