Replace ring_io with file_cache
This allows using the extra memory available on at32f435 which was not
possible with ring_io. It also has two side-goals of reducing the
diff and I/O behavior in image handlers compared to stable-v3 and
reducing overall code size.

While readahead was a core principal of ring_io, it is more of a bolt-on
for file_cache; file_cache focuses more on batched reads and writes.
File_cache tries to look more like the normal F_read, F_write, and
volume_cache with the main difference being that operations cannot span
sectors. file_cache_{read,write} could easily support multi-sector ops,
but no callers need such a thing as multi-sector reads/writes were for
batching.

file_cache still supports the "write directly to the cache" behavior
available in ring_io, but HFE and QD are the only images that make use
of it. The other images use a write buffer to avoid the read-then-write
that was required with ring_io. Those other images now do not do
reading-during-writing when their sector sizes are 512 or larger, which
matches stable-v3 behavior.

Since file_cache doesn't require read-before-write, the readahead can be
higher or lower priority than writes.