floppy: apply NCO-based 2nd-order PLL to WDATA line
Centurion Finch Floppy Controller
(FFC, https://github.com/Nakazoto/CenturionComputer/wiki/FFC-Board)
is an 8" floppy controller that applies up to 350ns write
precompensation. FlashFloppy's existing approach of re-aligning
the bit clock to the last pulse edge causes enough additional error
for bitcells to be missed or injected in this case.

This commit applies a 2nd-order (proportional and integral) PLL to
the incoming WDATA to mitigate both frequency offset and extreme
write precompensation. The implementation is based on a
virtual numerically-controlled oscillator nominally running at the
same frequency at the WDATA timer but with a 16.16 fixed-point
representation allowing for considerable precision in frequency
adjustment.  Proportional and integral constants were calcuated to
nominally provide a 1440kHz natural loop frequency and a damping
factor of 0.25 to allow for 2% frequency offset and strong jitter
rejection against write precompenstaion.