img: new geometry list for Otrona Attaché
This adds support for Otrona Attaché host computers.
A sample image is available on bitsavers:
http://www.bitsavers.org/bits/Otrona/Attache/Attache_Boot_2.2.5.IMD

The Attaché had an optional 98 TPI drive, but I don't have any way of
testing it so I only implemented the default 48 TPI setup.

Specs:
* NEC µPD765 controller
* Remex RFD480 drive
* 40 tracks
* 2 heads
* 10 sectors per track
* 512 byte sectors
* No skew

Note: The Otrona requires a RDY signal on pin 6, which is not connected
on the Gotek. A bodge is needed between pin 2 and 6, and pin 2 needs to
be configured as RDY.